Separate GPU transformation code out of DCE.
---
